Moves tokenizer to its own directory
