+++ src/ast/tests/tokenize.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "@jest/globals";
-import { operators } from "../airtable-formula-reference.json";
+import { operators } from "../../airtable-formula-reference.json";
 import { tokenize } from "../tokenize";
 
 describe("tokenize()", () => {
@@ -784,7 +784,7 @@ describe("tokenize()", () => {
             "value": "
              *  > 100,
             ,
-            
+
           ",
           },
         ],
